Make the tile iterator a class.
Closes #734.
